Use UnlockProfile component on MediaCard
