+++ packages/react-app/components/cards/MediaCard.jsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useContext, useEffect, useState } from "react";
 import { useColorModeValue } from "@chakra-ui/color-mode";
-import { Box, Flex, Stack, Heading, Code, HStack, VStack } from "@chakra-ui/layout";
+import { Box, Flex, Stack, Heading, Code, HStack } from "@chakra-ui/layout";
 import {
   Avatar,
   Popover,
@@ -15,18 +15,6 @@ import {
   Tag,
   TagLabel,
   SimpleGrid,
-  useToast,
-  useDisclosure,
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalHeader,
-  ModalCloseButton,
-  ModalBody,
-  ModalFooter,
-  FormLabel,
-  Input,
-  Checkbox,
 } from "@chakra-ui/react";
 import { Icon, EmailIcon, InfoIcon, PhoneIcon, LinkIcon } from "@chakra-ui/icons";
 import { BsFillPersonLinesFill } from "react-icons/bs";
@@ -34,12 +22,13 @@ import { GrLocation } from "react-icons/gr";
 import { ethers } from "ethers";
 import axios from "axios";
 import Blockies from "react-blockies";
-import { NETWORKS } from "../../constants";
+import { useRouter } from "next/dist/client/router";
+import UnlockProfile from "../UnlockProfile";
 
 import { Web3Context } from "../../helpers/Web3Context";
-import { useDebounce } from "../../hooks";
 
 function MediaCard({
+  did,
   account,
   avatarSrc,
   coverSrc,
@@ -51,10 +40,6 @@ function MediaCard({
   twitter,
   hasWebAccount,
   github,
-  primaryActionOnClick,
-  primaryAction,
-  secondaryActionOnClick,
-  secondaryAction,
   privateProfile,
   dRecruitContract,
   tokenContract,
@@ -62,16 +47,8 @@ function MediaCard({
 }) {
   const [decryptedData, setDecryptedData] = useState();
   const [canView, setCanView] = useState(false);
-  const [stakeAmount, setStakeAmount] = useState();
-  const debouncedStakeAmount = useDebounce(stakeAmount, 500);
-  const [approvalState, setApprovalState] = useState(); // ENOUGH || NOT_ENOUGH || LOADING
-  const [currAllowance, setCurrAllowance] = useState();
-  const [unlimitedAllowanceWanted, setUnlimitedAllowanceWanted] = useState(true);
   const context = useContext(Web3Context);
-
-  const toast = useToast();
-
-  const { isOpen, onOpen, onClose } = useDisclosure();
+  const router = useRouter();
 
   useEffect(() => {
     let isValidRecipient = false;
@@ -92,166 +69,8 @@ function MediaCard({
     })();
   }, [privateProfile]);
 
-  const handleRequestPrivateProfileUnlock = async () => {
-    if (!stakeAmount || +stakeAmount < +"0.1") {
-      return toast({ title: "Please enter a valid stake amount. Minimum 0.1 tokens.", status: "error" });
-    }
-    const weiStakeAmount = ethers.utils.parseEther(stakeAmount);
-    try {
-      // Only ask for allowance if it is not enough
-      if (approvalState === "NOT_ENOUGH") {
-        const tx = await tokenContract.approve(
-          dRecruitContract.address,
-          unlimitedAllowanceWanted ? ethers.constants.MaxUint256.toString() : weiStakeAmount,
-        );
-        toast({
-          title: "Approval transaction sent",
-          description: (
-            <text>
-              Your transaction was successfully sent{" "}
-              <a href={`${NETWORKS.mumbai.blockExplorer}/tx/${tx.hash}`} target="_blank" rel="noopener noreferrer">
-                <LinkIcon color="white" />
-              </a>
-            </text>
-          ),
-          status: "success",
-        });
-        await tx.wait();
-        toast({
-          title: "Approval transaction confirmed",
-          description: (
-            <text>
-              Your transaction was confirmed{" "}
-              <a href={`${NETWORKS.mumbai.blockExplorer}/tx/${tx.hash}`} target="_blank" rel="noopener noreferrer">
-                <LinkIcon color="white" />
-              </a>
-            </text>
-          ),
-          status: "success",
-        });
-      }
-      const tx = await dRecruitContract.request(privateProfile.tokenId, weiStakeAmount, {
-        value: 0,
-      });
-      toast({
-        title: "Request transaction sent",
-        description: (
-          <text>
-            Your transaction was successfully sent{" "}
-            <a href={`${NETWORKS.mumbai.blockExplorer}/tx/${tx.hash}`} target="_blank" rel="noopener noreferrer">
-              <LinkIcon color="white" />
-            </a>
-          </text>
-        ),
-        status: "success",
-      });
-      const receipt = await tx.wait();
-      toast({
-        title: "Request transaction confirmed",
-        description: (
-          <text>
-            Your transaction was confirmed{" "}
-            <a href={`${NETWORKS.mumbai.blockExplorer}/tx/${tx.hash}`} target="_blank" rel="noopener noreferrer">
-              <LinkIcon color="white" />
-            </a>
-          </text>
-        ),
-        status: "success",
-      });
-      onClose();
-    } catch (err) {
-      toast({
-        title: "Request transaction failed",
-        description: err.message + (err.data ? ` ${err.data.message}` : ""),
-        status: "error",
-      });
-    }
-    setStakeAmount();
-    // const { data } = await axios.get(`${process.env.NEXT_PUBLIC_API_URL}/unlock/${privateProfile.tokenId}`, {
-    //   withCredentials: true,
-    // });
-    // return data;
-  };
-
-  useEffect(() => {
-    async function exec() {
-      const weiStakeAmount = ethers.utils.parseEther(debouncedStakeAmount);
-      const allowance = await tokenContract.allowance(context.address, dRecruitContract.address);
-      if (allowance.lt(weiStakeAmount)) {
-        setApprovalState("NOT_ENOUGH");
-      } else {
-        setApprovalState("ENOUGH");
-      }
-    }
-    if (debouncedStakeAmount) {
-      setApprovalState("LOADING");
-      exec();
-    }
-  }, [isOpen, debouncedStakeAmount]);
-
-  useEffect(() => {
-    if (unlimitedAllowanceWanted) {
-      setApprovalState("NOT_ENOUGH");
-    }
-  }, [unlimitedAllowanceWanted]);
-
-  useEffect(() => {
-    async function exec() {
-      const allowance = await tokenContract.allowance(context.address, dRecruitContract.address);
-      setCurrAllowance(allowance);
-    }
-    exec();
-  }, [isOpen]);
-
   return (
     <>
-      <Modal isOpen={isOpen} onClose={onClose}>
-        <ModalOverlay />
-        <ModalContent>
-          <ModalHeader>Request unlock information</ModalHeader>
-          <ModalCloseButton />
-          <ModalBody>
-            <VStack justifyItems="left">
-              <FormLabel htmlFor="stakeAmount">Enter stake amount in {tokenMetadata && tokenMetadata.symbol}</FormLabel>
-              <Input
-                value={stakeAmount}
-                onChange={e => setStakeAmount(e.target.value)}
-                placeholder="Minimum 0.1"
-                borderColor="purple.500"
-              />
-
-              {currAllowance && currAllowance.lt(ethers.constants.MaxUint256) && (
-                <HStack>
-                  <Checkbox defaultIsChecked
-                    value={unlimitedAllowanceWanted}
-                    onChange={e => setUnlimitedAllowanceWanted(e.target.checked)}
-                  >
-                    Approve with unlimited allowance?
-                  </Checkbox>
-                </HStack>
-              )}
-            </VStack>
-          </ModalBody>
-
-          <ModalFooter>
-            <Button
-              width="100%"
-              colorScheme="blue"
-              onClick={() => {
-                handleRequestPrivateProfileUnlock();
-              }}
-              disabled={!approvalState || approvalState === "LOADING"}
-              isLoading={approvalState === "LOADING"}
-            >
-              {!approvalState || approvalState === "LOADING"
-                ? "Enter amount"
-                : approvalState === "ENOUGH"
-                ? "Confirm"
-                : "Approve"}
-            </Button>
-          </ModalFooter>
-        </ModalContent>
-      </Modal>
       <Box
         maxW={"350px"}
         w={"full"}
@@ -288,7 +107,9 @@ function MediaCard({
         <Box p={6}>
           <Stack spacing={0} align="left" mb={5}>
             <Heading fontSize={"2xl"} fontWeight={600} fontFamily={"body"}>
-              {heading || "Anonymous"}
+              <a href="#" onClick={() => router.push("/profile/" + did)}>
+                {heading || "Anonymous"}
+              </a>
             </Heading>
             <Heading fontSize={"lg"} fontWeight={500} fontFamily={"body"}>
               {description}
@@ -345,15 +166,13 @@ function MediaCard({
               </Stack>
             </Box>
           )}
-          <Button
-            disabled={canView}
-            w={"full"}
-            mt={8}
-            // TODO: get dev main address
-            onClick={onOpen}
-          >
-            {canView ? "✔️ Information already unlocked" : primaryAction}
-          </Button>
+          <UnlockProfile
+            canView={canView}
+            dRecruitContract={dRecruitContract}
+            tokenContract={tokenContract}
+            privateProfile={privateProfile}
+            tokenMetadata={tokenMetadata}
+          />
         </Box>
       </Box>
     </>
